feat: implement barebones cli web media player
been falling into some web media rabbit holes using my browser and
i'd like to regain some control over that by playing only the media
i'm deliberately looking for using the the command line. it simply
does some search for 1 page of results and plays a selected result
using `mpv`.

this is a re-implementation of an older tool i had cooked last year
that worked more like a media server, but that i stopped used after
some time. i stripped all the wrapping server logic and wrapped it
into a process object.

unfortunately my `mpv` is currently broken and doesn't work the same
as an older version of this tool used to work. i'm checking it in
as is in hopes I can get mpv working again... in the meantime i will
probably extract the mpv component and replace it with a subprocess
running `yt-dlp` and `ffplay` which seems to work, but doesn't seem
to provide socket control as `mpv` does.

```
yt-dlp -f mp4 "$url" -o - | ffplay - -autoexit -loglevel quiet
```