feat: implement tool to automatically ocr images in directory
moved over an old script to ocr images using macos vision framework
and paired it with a process that watches for images created in a
directory to automatically use ocr to extract text and save it on
a text file with a matching filename (sans extension ofc).

wanted to have a way to carry the swift script around w/out having
to sort out paths for static assets so i shoved it into a multiline
string in the python module itself. if `scribe` cannot find the
script where it expects it to be then it will simply create it.

needs some logging and a few other niceties... like proper error
handling.