feat: implement graphics capabilities
playing around with graphics again, added a bunch of facilities to
add window and graphics functionalities to some of the programs in
here. it uses sdl2 as the underlying library - and pillow to load
images - but the rest is done as cpu graphics operations.

also added click support to have commands and subcommands.