feat: load images from filesystem for graphics lib
also, discovered `draw_image` is broken as the bound calculation
is incorrect and it's only rendering half the width of the source
image.

i'm considering replacing a lot of this drawing code with my the
old `bitblt` code i ported from the smalltalk blue book a couple
of years back. that code wraps all these edge cases and opens the
door to port all the graphics code in the bluebook as well.