broken: sketch port of oo bitblt into procedural
took an older port of ST80's bitblt and started porting it (again) to
a more procedural implementation. it's not quite working at the moment
as it is missing some pieces. i'll shape it up a bit more and - maybe
- will merge it into the `Form` object to make it more convenient to
use.

interestingly this will likely change the font rendering logic if i
decide to use bitblt for drawing glyphs.

there's a lot of changes i have planned for the windowing / gui stuff
that will depend a lot on bitblt so i really need to go back and get
this working. the code looks mostly ok, but it's untested and likely
has bugs... it's also missing functionality.