feat: complete nabu and set up app installer
wrapped a few pending items for `nabu` (may change name btw) though
there is still work to be done on the gui/draw module to refactor all
the windowing stuff.

also decided to find a way to package some of these tools as apps and
after looking around pyinstaller seemed to be the thing. i still have
to create some icons to get apps out of this, but tbh i may scrap this
at any point... we will see.