fix: add starting point to lines
had missed this, it's not optional as the smalltalk way... withouth it
things like rectangle drawings look incomplete.

noticed two new things:
- the line start is not tracking the real x,y position and I am not
  fully sure why that is.
- during the port i did some renaming and totally botched the line
  drawing implementation, instead of `copy_bits` it ought to be
  calling `copy_loop` which in this version is the `bitblt` method.

when i tried to fix the second issue, found that clipping is leading
to some artifacts in line drawing primitives. artifacts that are not
present on my first more smalltalk-ish implementation of doodle. need
to look over that code again.