fix: translate mouse coordinates in canvas form.
from previous commit:

> - the line start is not tracking the real x,y position and I am not
>   fully sure why that is.

well it was because the canvas' corner is not at 0, 0 so the mouse
coordinates are off by the top and left margin on the canvas (TODO
need to make those constants or parameters). the fix was simply to
translate the coordinates to use on the call to `draw_line`.

this - and other code - would be cleaner with more functionality in
`Point` class... translate, add, sub, mul, div, dot_product, etc.

and this would not have come up if `Form` had an offset attribute like
the one from my earlier implementations. this offset is used in line
drawing when computing the destination point. this bug wouldn't have
even come up. maybe i'll add that later.