feat: improve font support
implemented a way to load hex fonts into the `Font` object and added a
manager for - potentially - multiple fonts.

the idea is that by default one can create a `~/hexfonts` directory
and drop a bunch of hexfonts with a specific name format and then let
the manager load them from there OR one can define a custom path font
manager and pass it to the window when instantiating it.

for now I turned the one font i had to hex format with a few lines of
python. to set this up one copies `fonts/*` to `~/hexfonts` and be
done. alternatively one could copy the hex format of unscii or unifont
there as well.

working now on a simple script to take bdf fonts and convert them to
hexformat. though i suppose i could also support `load_bdf_font`.