feat: add support for bdf fonts
rather than convert bdf fonts into hexfonts it seemed more sensible to
add support for bdf fonts to the font manager. after all it's the same
code - maybe less so - and it'd be one fewer step to get new fonts
into the system.

now one just needs to download a bdf font into `~/wastefonts` and be
done. it can either be a hexfont or a bdf font.

added some commands to doodle for the sole purpose of testing this
out.

next up is to add new graphic primitives:
- draw_circle
- fill_rect
- fill_circle
- draw_path
- draw_arc
- fill_polygon?

and then some new graphics objects
- pen
- path
- arc

after that the name of the game is to add a command interpreter to
every window.

this is morphing already into what was meant to be my end goal? maybe
I should move hard in that direction now?