fix: draw glyphs wider than 8 pixels properly
as i started trying slightly larger fonts i ran into an issue when
they were getting cut off quite a bit. looking further into it i
noticed that so far it had been working ok as all fonts i was using
were 8 pixels wide so the right shift by the width was working great,
and it would likely work for multiples of 8, but was breaking for
10x20, 12x24, and so on...

added some simple logic to use the right bitwidth when right shifting.

also, added a fallback glyph to return when the font doesn't have one
for the codepoint. it's ugly as this is a specific's font glyph and so
it looks odd in other fonts. it'll be nicer if this is generated based
on the actual font's dimension... hmmm?

fount a bunch of fonts on gh to download to `~/wastefonts` to try
out. i particularly liked the dec vt220 ones (10x20 size is quite
nice)
[here](https://github.com/htayj/DEC-Fonts/tree/main/dist/fonts/bdf)

there's also
- [bitmap-fonts](https://github.com/Tecate/bitmap-fonts/tree/master?tab=readme-ov-file)
- [bdf-fonts](https://github.com/szilard10/bdf-fonts)
- [spleen](https://github.com/fcambus/spleen/tree/master)
- [cnxt](https://github.com/cbytensky/cnxt/blob/main/cnxt.bdf)