test: implement mechanism to test gui tools
tbh it feels strange to unit test these gui tools, but i want to at
least know if some change in `gui` or `draw` crashes the tool. it's
much harder to test that it behaves as expected fully, specially for
rendering... but i'll take what i can get.

also this feels kinda hacky and yet kinda neat :grin: ... tho i don't
like the idea of having to subclass the thing. maybe i'll just define
a method to patch the object as i do in the subclassed definition.

this is not what i set out to do... but once i saw a bunch of 0%
coverage modules my neurosis got the better of me lol. not that code
coverage is a great proxy for working software - specially on python -
but having these little tests will be useful for something and it's
not like it takes a ton of effort... and i learned a bit about
pushing synthetic SDL2 events.