feat: start implementation of a toy not quite a lisp lang
defines some starting types

implements a character stream and a reader that is able to read
strings and not much else.

this very loosely follows the Make A Lisp (MAL) guide and borrows some things
from my not-extensive experience with Common Lisp, Clojure, and Scheme.

i've done a few write a lisp tutorials in the past, but always stop
with it once it's done and don't do much else. i want to make this a
bit more useful for some experiments i have in mind.

while this is a toy, it is a toy that i intend to use to play and
understand some nonsense that's been in my head for a while now.

i didn't really want to make a lisp-like thing and was going to play
with an extended version of `lox` from Crafting Interpreters, but i
couldn't argue with some of the "ease" of parsing and manipulating
lists lol. i will still go back to `lox` and extend it further... i'll
have to if i get to the point where i can move some of these ideas
into a bytecode vm; if time ever permits.