fix: reset previous mouse coordinates to prevent stray points
been experimenting drawing in `doodle` with an ipad mini + apple
pencil using it as an external display... it's much nicer than drawing
with a trackpad or mouse.

however, there was a small bug that i kept running into occasionally
with the mouse that happens every time w/ the pencil due to the fact
that there's no mouse motion event w/out mouse pressed on a touch
surface.

with the mouse sometimes the mouse device previous position would not
update before the mouse was pressed and `doodle` would draw a line
from the stray previous position to the new click position.

with the pencil this happens every time one lifts the pencil to start
a new line in a different section of the drawing surface. one ends up
with all drawings connected by lines between their last point of the
last drawing to the first point of the new drawing.

i just implemented a quick fix by splitting the mouse handlers to
differentiate between the 3 events (mouse pressed, mouse released,
and mouse moved) and updated `doodle` to reset the mouse previous
position during mouse pressed.