refactor: remove dependency on yt-dlp library and ffplay
renamed media module and plugged into melody to replace the `ffplay`
process and the use of `yt-dlp` as a library (ofc `mpv` is still using
it internally, but now i don't have to deal with it myself).

the main reason i wanted `mpv` is that it has a server mode that one
can communicate via a unix socket using its command interface. this
let's me start run the `mpv` process in the background and connect to
it using a unix socket, then send commands directly from the ui
process which can access a reference to that socket. this way i can
pause/resume/stop the media playback through the ui instead of using
two application windows as i did before with `ffplay`.Granted there
may be a way to do the same there, but i didn't know and i initially
wanted to use `mpv` so :shrug:

the command interface also supports querying so i can get back
playback information and use it to render it tho i yet don't implement
it.